Renames app to 'banksaurus'
