+++ cmd/banksaurus/commands/command_report.go
@@ -3,11 +3,11 @@ package commands
 import (
 	"os"
 
-	"github.com/luistm/go-bank-cli/bank/transactions"
-	"github.com/luistm/go-bank-cli/cmd/bankcli/configurations"
-	"github.com/luistm/go-bank-cli/infrastructure/csv"
-	"github.com/luistm/go-bank-cli/infrastructure/sqlite"
-	"github.com/luistm/go-bank-cli/lib/sellers"
+	"github.com/luistm/banksaurus/bank/transactions"
+	"github.com/luistm/banksaurus/cmd/banksaurus/configurations"
+	"github.com/luistm/banksaurus/infrastructure/csv"
+	"github.com/luistm/banksaurus/infrastructure/sqlite"
+	"github.com/luistm/banksaurus/lib/sellers"
 )
 
 // Report handles reports
