+++ cmd/banksaurus/commands/command_seller.go
@@ -3,9 +3,9 @@ package commands
 import (
 	"os"
 
-	"github.com/luistm/go-bank-cli/cmd/bankcli/configurations"
-	"github.com/luistm/go-bank-cli/infrastructure/sqlite"
-	"github.com/luistm/go-bank-cli/lib/sellers"
+	"github.com/luistm/banksaurus/cmd/banksaurus/configurations"
+	"github.com/luistm/banksaurus/infrastructure/sqlite"
+	"github.com/luistm/banksaurus/lib/sellers"
 )
 
 // Seller commands
