+++ infrastructure/infrastructure.go
@@ -3,8 +3,8 @@ package infrastructure
 import (
 	"io"
 
-	"github.com/luistm/go-bank-cli/bank/transactions"
-	"github.com/luistm/go-bank-cli/lib"
+	"github.com/luistm/banksaurus/bank/transactions"
+	"github.com/luistm/banksaurus/lib"
 )
 
 // SQLStorage for handling SQL databases
