+++ lib/sellers/interactors.go
@@ -1,8 +1,8 @@
 package sellers
 
 import (
-	"github.com/luistm/go-bank-cli/lib"
-	"github.com/luistm/go-bank-cli/lib/customerrors"
+	"github.com/luistm/banksaurus/lib"
+	"github.com/luistm/banksaurus/lib/customerrors"
 )
 
 // NewInteractor creates a new Interactor object for sellers
