+++ lib/sellers/interactors_test.go
@@ -5,10 +5,10 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/luistm/go-bank-cli/elib/testkit"
+	"github.com/luistm/banksaurus/elib/testkit"
 
-	"github.com/luistm/go-bank-cli/lib"
-	"github.com/luistm/go-bank-cli/lib/customerrors"
+	"github.com/luistm/banksaurus/lib"
+	"github.com/luistm/banksaurus/lib/customerrors"
 )
 
 func TestUnitInteractorCreate(t *testing.T) {
